Block test in production

See merge request eip/peer!56

+++ server/src/database.ts
@@ -61,7 +61,11 @@ export default class Database {
      * @constructor - default constructor.
      */
   static async DatabaseImport(qf: pgp.QueryFile) {
-    await this.db.any(qf);
+    if (process.env.NODE_ENV === "production") {
+      throw Error("You are running in production!");
+    } else {
+      await this.db.any(qf);
+    }
   }
 
     /**
@@ -70,8 +74,12 @@ export default class Database {
      * @constructor - default constructor.
      */
     static async DatabaseDrop() {
-      await this.db.any("DROP SCHEMA IF EXISTS public CASCADE");
-      await this.db.any("CREATE SCHEMA public");
+      if (process.env.NODE_ENV === "production") {
+        throw Error("You are running in production!");
+      } else {
+        await this.db.any("DROP SCHEMA IF EXISTS public CASCADE");
+        await this.db.any("CREATE SCHEMA public");
+      }
     }
 
     /**
